+++ main.go
@@ -30,10 +30,10 @@ import (
 	"github.com/spf13/pflag"
 	"google.golang.org/grpc"
 
+	grpcApi "github.com/awfm9/flow-dps/api/grpc"
 	"github.com/awfm9/flow-dps/api/rest"
 	"github.com/awfm9/flow-dps/chain"
 	"github.com/awfm9/flow-dps/feeder"
-	grpcApi "github.com/awfm9/flow-dps/grpc"
 	"github.com/awfm9/flow-dps/mapper"
 	"github.com/awfm9/flow-dps/state"
 )
